sample C++ interop

+++ library/setup.py
@@ -1,8 +1,9 @@
+from pathlib import Path
 from urllib.request import urlretrieve
 from zipfile import ZipFile
 from setuptools import setup, find_namespace_packages
 
-from data_outrigger.utils.file_utils import regenerate_dir, DROW_WEIGHTS_PATH
+from pybind11.setup_helpers import Pybind11Extension, build_ext
 
 VERSION = "0.0.1"
 DROW_DATA = "https://github.com/VisualComputingInstitute/DROW/releases/download/v2/DROWv2-data.zip"
@@ -18,13 +19,22 @@ dependencies = [
     "matplotlib~=3.7"
 ]
 
+ext_modules = [
+    Pybind11Extension("cpp_binding",
+        ["cpp_core/main.cpp"],
+        define_macros = [("VERSION_INFO", VERSION)],
+    ),
+]
 
 
-include_dir = regenerate_dir("data_outrigger/include")
-file, _ = urlretrieve(DROW_DATA)
-with ZipFile(file, "r") as archive:
-    archive.extractall(include_dir)
-urlretrieve(DROW_WEIGHTS, include_dir / DROW_WEIGHTS_PATH.name)
+
+include_dir = Path("data_outrigger/include")
+if not include_dir.exists():
+    include_dir.mkdir(parents=True)
+    file, _ = urlretrieve(DROW_DATA)
+    with ZipFile(file, "r") as archive:
+        archive.extractall(include_dir)
+    urlretrieve(DROW_WEIGHTS, include_dir / Path("weights.pth.tar"))
 
 setup(
     name="data_outrigger",
@@ -32,10 +42,11 @@ setup(
     description="Library for all simple and commonly used ROS utilities I use.",
     author="Aleksandr Sergeev",
     author_email="alexander.sergeev@onmail.com",
-    url='http://github.com/pseusys/...',
+    url='http://github.com/pseusys/follow_the_drow',
     packages=find_namespace_packages(),
+    ext_modules=ext_modules,
+    cmdclass={"build_ext": build_ext},
     install_requires=dependencies,
     include_package_data=True,
+    python_requires=">=3.9",
 )
-
-# TODO: finish url
